Rename InternalStatistics to InternalStatisticsReceiver
