+++ stroom-core-server/src/main/java/stroom/spring/MetaDataStatisticConfiguration.java
@@ -4,7 +4,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import stroom.internalstatistics.MetaDataStatisticImpl;
 import stroom.internalstatistics.MetaDataStatisticTemplate;
-import stroom.statistics.internal.InternalStatistics;
+import stroom.statistics.internal.InternalStatisticsReceiver;
 
 import java.util.Arrays;
 
@@ -14,8 +14,8 @@ public class MetaDataStatisticConfiguration {
      * This bean must be returned as a class and not an interface otherwise annotation scanning will not work.
      */
     @Bean
-    public MetaDataStatisticImpl metaDataStatistic(final InternalStatistics internalStatistics) {
-        final MetaDataStatisticImpl metaDataStatistic = new MetaDataStatisticImpl(internalStatistics);
+    public MetaDataStatisticImpl metaDataStatistic(final InternalStatisticsReceiver internalStatisticsReceiver) {
+        final MetaDataStatisticImpl metaDataStatistic = new MetaDataStatisticImpl(internalStatisticsReceiver);
         metaDataStatistic.setTemplates(Arrays.asList(
                 new MetaDataStatisticTemplate(
                         "Meta Data-Streams Received",
