+++ stroom-core-server/src/test/java/stroom/volume/server/TestVolumeServiceImpl.java
@@ -33,7 +33,7 @@ import stroom.node.shared.Rack;
 import stroom.node.shared.Volume;
 import stroom.node.shared.Volume.VolumeType;
 import stroom.node.shared.VolumeState;
-import stroom.statistics.internal.InternalStatistics;
+import stroom.statistics.internal.InternalStatisticsReceiver;
 import stroom.util.config.StroomProperties;
 import stroom.util.io.FileUtil;
 import stroom.util.spring.StroomBeanStore;
@@ -207,8 +207,8 @@ public class TestVolumeServiceImpl extends StroomUnitTest {
 
         public MockVolumeService(final StroomEntityManager stroomEntityManager, final NodeCache nodeCache,
                                  final StroomPropertyService stroomPropertyService, final StroomBeanStore stroomBeanStore,
-                                 final InternalStatistics internalStatistics) {
-            super(stroomEntityManager, nodeCache, stroomPropertyService, stroomBeanStore, internalStatistics);
+                                 final InternalStatisticsReceiver internalStatisticsReceiver) {
+            super(stroomEntityManager, nodeCache, stroomPropertyService, stroomBeanStore, internalStatisticsReceiver);
         }
 
         @Override
